refactor: use flutter hooks fully
