+++ lib/features/bottom_scroll_sheet/sliver_multi_tabber_builder.dart
@@ -6,40 +6,21 @@ import "../../config/map_view_config.dart";
 import "../../config/ui_config.dart";
 import "../../theme/app_theme.dart";
 import "../../utils/context_extensions.dart";
-import "../map_view/controllers/bottom_sheet_controller.dart";
 import "../multilayer_map/data/model/multilayer_item.dart";
 import "multilayer_map_single_entity_list.dart";
+import "use_tab_scroll_sync.dart";
 
 /// Helper widget to build slivers for multi-tab view
-class SliverMultiTabberBuilder extends StatefulWidget {
+class SliverMultiTabberBuilder extends HookWidget {
   const SliverMultiTabberBuilder({required this.tabs, required this.scrollController});
 
   final List<({String title, MultilayerMapSingleEntityList<MultilayerItem> Function() builder})> tabs;
   final ScrollController scrollController;
 
   @override
-  State<SliverMultiTabberBuilder> createState() => _SliverMultiTabberBuilderState();
-}
-
-class _SliverMultiTabberBuilderState extends State<SliverMultiTabberBuilder> {
-  late final List<GlobalKey> sectionKeys;
-  late final ValueNotifier<int> selectedTabIndex;
-
-  @override
-  void initState() {
-    super.initState();
-    sectionKeys = List.generate(widget.tabs.length, (_) => GlobalKey());
-    selectedTabIndex = ValueNotifier(0);
-  }
-
-  @override
-  void dispose() {
-    selectedTabIndex.dispose();
-    super.dispose();
-  }
-
-  @override
   Widget build(BuildContext context) {
+    final sectionKeys = useMemoized(() => List.generate(tabs.length, (_) => GlobalKey()), [tabs.length]);
+
     return SliverMainAxisGroup(
       slivers: [
         SliverPersistentHeader(
@@ -51,12 +32,7 @@ class _SliverMultiTabberBuilderState extends State<SliverMultiTabberBuilder> {
                 alignment: Alignment.centerLeft,
                 child: SizedBox(
                   height: context.textScaler.clamp(maxScaleFactor: 2).scale(40),
-                  child: _TabBarWidget(
-                    tabs: widget.tabs,
-                    selectedTabIndex: selectedTabIndex,
-                    scrollController: widget.scrollController,
-                    sectionKeys: sectionKeys,
-                  ),
+                  child: _TabBarWidget(tabs: tabs, scrollController: scrollController, sectionKeys: sectionKeys),
                 ),
               ),
             ),
@@ -64,7 +40,7 @@ class _SliverMultiTabberBuilderState extends State<SliverMultiTabberBuilder> {
           ),
         ),
         // Add each tab's content
-        for (var i = 0; i < widget.tabs.length; i++)
+        for (var i = 0; i < tabs.length; i++)
           SliverMainAxisGroup(
             slivers: [
               // Add a trackable marker widget at the start of each section
@@ -78,7 +54,7 @@ class _SliverMultiTabberBuilderState extends State<SliverMultiTabberBuilder> {
                         horizontal: MapViewBottomSheetConfig.horizontalPadding,
                         vertical: NavigationTabViewConfig.smallerPadding,
                       ),
-                sliver: widget.tabs[i].builder(),
+                sliver: tabs[i].builder(),
               ),
             ],
           ),
@@ -89,91 +65,27 @@ class _SliverMultiTabberBuilderState extends State<SliverMultiTabberBuilder> {
 
 /// Widget for the tab bar
 class _TabBarWidget extends HookConsumerWidget {
-  const _TabBarWidget({
-    required this.tabs,
-    required this.selectedTabIndex,
-    required this.scrollController,
-    required this.sectionKeys,
-  });
+  const _TabBarWidget({required this.tabs, required this.scrollController, required this.sectionKeys});
 
   final List<({String title, MultilayerMapSingleEntityList<MultilayerItem> Function() builder})> tabs;
-  final ValueNotifier<int> selectedTabIndex;
   final ScrollController scrollController;
   final List<GlobalKey> sectionKeys;
 
   @override
   Widget build(BuildContext context, WidgetRef ref) {
     final tabController = useTabController(initialLength: tabs.length);
-    final isScrollingToTab = useRef(false);
-
-    // Listen to scroll to update active tab
-    useEffect(() {
-      void onScroll() {
-        if (isScrollingToTab.value) return;
-
-        // Find the first section that's visible near the top
-        for (var i = sectionKeys.length - 1; i >= 0; i--) {
-          final ctx = sectionKeys[i].currentContext;
-          if (ctx == null) continue;
-
-          final renderBox = ctx.findRenderObject() as RenderBox?;
-          if (renderBox == null || !renderBox.attached || !renderBox.hasSize) continue;
-
-          final position = renderBox.localToGlobal(Offset.zero);
-
-          // If this section's marker is at or above 200px from top, it's active
-          if (position.dy <= 200) {
-            if (selectedTabIndex.value != i) {
-              selectedTabIndex.value = i;
-              if (tabController.index != i) {
-                tabController.animateTo(i);
-              }
-            }
-            break;
-          }
-        }
-      }
-
-      scrollController.addListener(onScroll);
-      return () => scrollController.removeListener(onScroll);
-    }, [scrollController]);
-
-    // Sync tab controller with selected index
-    useEffect(() {
-      void listener() {
-        if (tabController.index != selectedTabIndex.value) {
-          tabController.animateTo(selectedTabIndex.value);
-        }
-      }
-
-      selectedTabIndex.addListener(listener);
-      return () => selectedTabIndex.removeListener(listener);
-    }, []);
+    final (:selectedTabIndex, :onTabTap) = useTabScrollSync(
+      sectionKeys: sectionKeys,
+      scrollController: scrollController,
+      tabController: tabController,
+      ref: ref,
+    );
 
     return Theme(
       data: Theme.of(context).copyWith(splashColor: Colors.transparent, highlightColor: Colors.transparent),
       child: TabBar(
         controller: tabController,
-        onTap: (index) async {
-          final ctx = sectionKeys[index].currentContext;
-          if (ctx != null) {
-            isScrollingToTab.value = true;
-            selectedTabIndex.value = index;
-
-            // First expand the bottom sheet fully
-            await ref.watch(bottomSheetPixelsProvider.notifier).onSearchBoxTap();
-            if (ctx.mounted) {
-              await Scrollable.ensureVisible(ctx, duration: const Duration(milliseconds: 300), curve: Curves.easeInOut);
-              await Future<void>.delayed(const Duration(milliseconds: 50));
-              if (ctx.mounted) {
-                // ! this is nasty hack due to some bug with the `Scrollable.ensureVisible` overshooting while scrolling to the tab
-                await Scrollable.ensureVisible(ctx, alignment: 0.08);
-                await Future<void>.delayed(const Duration(milliseconds: 200));
-              }
-            }
-            isScrollingToTab.value = false;
-          }
-        },
+        onTap: onTabTap,
         isScrollable: true,
         padding: EdgeInsets.zero,
         indicatorColor: Colors.transparent,
@@ -181,28 +93,25 @@ class _TabBarWidget extends HookConsumerWidget {
         tabAlignment: TabAlignment.start,
         tabs: List.generate(
           tabs.length,
-          (index) => ValueListenableBuilder(
+          (index) => Container(
             key: ValueKey("tab_$index"),
-            valueListenable: selectedTabIndex,
-            builder: (context, selectedIndex, child) => Container(
-              margin: EdgeInsets.only(bottom: NavigationTabViewConfig.smallerPadding * 1.5, left: index == 0 ? 8 : 0),
-              padding: const EdgeInsets.symmetric(
-                horizontal: NavigationTabViewConfig.universalPadding,
-                vertical: NavigationTabViewConfig.smallerPadding,
-              ),
-              decoration: BoxDecoration(
-                color: index == selectedIndex
-                    ? context.colorTheme.orangePomegranadeLighter
-                    : context.colorTheme.greyLight,
-                borderRadius: BorderRadius.circular(NavigationTabViewConfig.radius),
-              ),
-              child: Text(
-                tabs[index].title,
-                textAlign: TextAlign.center,
-                style: TextStyle(
-                  fontWeight: index == selectedIndex ? FontWeight.bold : FontWeight.normal,
-                  color: index == selectedIndex ? Colors.white : Colors.black,
-                ),
+            margin: EdgeInsets.only(bottom: NavigationTabViewConfig.smallerPadding * 1.5, left: index == 0 ? 8 : 0),
+            padding: const EdgeInsets.symmetric(
+              horizontal: NavigationTabViewConfig.universalPadding,
+              vertical: NavigationTabViewConfig.smallerPadding,
+            ),
+            decoration: BoxDecoration(
+              color: index == selectedTabIndex.value
+                  ? context.colorTheme.orangePomegranadeLighter
+                  : context.colorTheme.greyLight,
+              borderRadius: BorderRadius.circular(NavigationTabViewConfig.radius),
+            ),
+            child: Text(
+              tabs[index].title,
+              textAlign: TextAlign.center,
+              style: TextStyle(
+                fontWeight: index == selectedTabIndex.value ? FontWeight.bold : FontWeight.normal,
+                color: index == selectedTabIndex.value ? Colors.white : Colors.black,
               ),
             ),
           ),
